split up fluxes into another abstract fir class
