+++ docs/installation/source.rst
@@ -59,7 +59,7 @@ For unit tests to pass you will also need the following optional requirements:
 .. code-block:: bash
 
     pip install pynufft
-    pip install pylops==1.11.1
+    pip install pylops==2.3.1
 
 If you are using a ``conda`` environment, add the source repository as follows:
